fix(antlr-rust): fix a warning of int comparison
/home/sec3/x-ray/code-parser/src/antlr-rust/RustParser.h:343:49: warning: comparison of integers of different signs: 'size_t' (aka 'unsigned long') and 'char' [-Wsign-compare]
  bool next(char expect) { return _input->LA(1) == expect; }
                                  ~~~~~~~~~~~~~ ^  ~~~~~~

Note that the codes (`RustLexer.h` and `RustParser.h`) were initially
generated via ANTLR.

+++ code-parser/src/antlr-rust/RustLexer.h
@@ -160,7 +160,9 @@ public:
     return next == 0 || next == antlr4::Token::EOF;
   }
 
-  bool next(char expect) { return _input->LA(1) == expect; }
+  bool next(char expect) {
+    return _input->LA(1) == static_cast<size_t>(expect);
+  }
 
   bool floatDotPossible() {
     size_t next = _input->LA(1);
